Current complete GameLoop with relatively significant infrastructure.
Running main in PyCharm should produce a small window with a reasonable pulse of purple, and should end the PyCharm run command when window is closed with X button.

Unable to convert the keep_game_running loop into a named method at this time, but should be able to condense things as we move into event handling.